- data processing refactoring for the event card
