+++ src/utils/getTime.js
@@ -1,0 +1,7 @@
+import dayjs from 'dayjs'
+import dayjsUkrLocale from 'dayjs/locale/uk'
+
+
+export function getTime(time) {
+    return dayjs(time).locale(dayjsUkrLocale).format('HH:mm')
+}
