+++ src/views/application/events/my-events.vue
@@ -161,6 +161,9 @@ import { API } from '../../../workers/api-worker/api.worker'
 import { ROUTES } from '../../../router/router.const'
 import { PaginationWorker } from '../../../workers/pagination-worker'
 import { FilterPatch } from '../../../workers/api-worker/http/filter/filter.patch'
+import { addMinutes } from '../../../utils/addMinutes'
+import { getDate } from '../../../utils/getDate'
+import { getTime } from '../../../utils/getTime'
 
 import CONSTANTS from '../../../consts/index'
 
@@ -257,28 +260,7 @@ export default {
       router.push(ROUTES.APPLICATION.EVENTS.absolute)
     }
 
-    function getDate(date) {
-      return dayjs(date).locale(dayjsUkrLocale).
-      format(Number(dayjs(date).locale(dayjsUkrLocale).format('YYYY')) === new Date().getFullYear()
-      ? 'D MMMM'
-      : ' D MMMM, YYYY')
-    }
-
-    function getTime(time) {
-      return dayjs(time).locale(dayjsUkrLocale).format('HH:mm')
-    }
-
-    function addMinutes(time, minutesToAdd) {
-      let timeArray = time.split(':');
-      let hours = timeArray[0];
-      let originalMinutes = timeArray[1];
-      let date = new Date();
-      date.setHours(hours);
-      date.setMinutes(originalMinutes);
-      date.setMinutes(date.getMinutes() + minutesToAdd);
-      return date.toTimeString().substr(0, 5);
-    }
-
+  
     function handlingIncomeData(item) {
       return {
         ...item,
